Migration fix

+++ database/migrations/2024_01_02_063637_create_transfer_history_table.php
@@ -37,8 +37,11 @@ return new class () extends Migration {
     public function down(): void
     {
         Schema::dropIfExists('ownership_transfers');
-        Schema::table('products', function (Blueprint $table): void {
-            $table->dropColumn('visibility');
-        });
+
+        if (Schema::hasTable('prosucts')) {
+            Schema::table('products', function (Blueprint $table): void {
+                $table->dropColumn('visibility');
+            });
+        }
     }
 };
